requirements.txt
tensorflow
overpy
pandas
numpy
fastapi[standard]
starlette
scikit-learn
uvicorn
google-generativeai
pillow
matplotlib
